refactor: simpilify the ts-loader config
