+++ example/components/tsx-lang.vue
